Only use the system default zshenv if the user zshenv doesn't exist
Rather than setting the system default zshenv content and also the
content from /etc, load the /etc content and fall back to the system
content if it doesn't exist.

The purpose of this change is to allow a user to keep their own
default values without having the system force the user to unset or
change values back if the user already has custom changes.